+++ components/Header.tsx
@@ -4,6 +4,7 @@ import { getMixpanelUserContext, trackingEvents } from 'analytics/analytics'
 import { ContextConnected } from 'blockchain/network'
 import { AppLink } from 'components/Links'
 import { ConnectWalletButton } from 'components/navigation/content/ConnectWalletButton'
+import { WalletPanelMobile } from 'components/navigation/content/WalletPanelMobile'
 import { LANDING_PILLS } from 'content/landing'
 import { DISCOVER_URL } from 'features/discover/helpers'
 import { getUnreadNotificationCount } from 'features/notifications/helpers'
@@ -32,9 +33,7 @@ import { Box, Button, Card, Container, Flex, Grid, Image, SxStyleProp, Text } fr
 import { useOnMobile } from 'theme/useBreakpointIndex'
 
 import { useAppContext } from './AppContextProvider'
-import { MobileSidePanelPortal, ModalCloseIcon } from './Modal'
 import { NotificationsIconButton } from './notifications/NotificationsIconButton'
-import { useSharedUI } from './SharedUIProvider'
 import { UniswapWidgetShowHide } from './uniswapWidget/UniswapWidget'
 
 function Logo({ sx }: { sx?: SxStyleProp }) {
@@ -334,90 +333,6 @@ function UserDesktopMenu() {
   )
 }
 
-function MobileSettings() {
-  const { vaultFormToggleTitle, setVaultFormOpened } = useSharedUI()
-  const [opened, setOpened] = useState(false)
-  const { accountData$, context$, web3Context$ } = useAppContext()
-  const [context] = useObservable(context$)
-  const [accountData] = useObservable(accountData$)
-  const [web3Context] = useObservable(web3Context$)
-  const componentRef = useOutsideElementClickHandler(() => setOpened(false))
-
-  const shouldHideSettings = getShouldHideHeaderSettings(context, accountData, web3Context)
-
-  if (shouldHideSettings) return null
-
-  return (
-    <>
-      <Flex
-        sx={{
-          position: 'fixed',
-          bottom: 0,
-          left: 0,
-          right: 0,
-          bg: 'rgba(255,255,255,0.9)',
-          p: 3,
-          justifyContent: 'space-between',
-          gap: 2,
-          zIndex: 3,
-        }}
-      >
-        <Button
-          variant="menuButton"
-          onClick={() => setOpened(true)}
-          sx={{ p: 1, width: vaultFormToggleTitle ? undefined : '100%', color: 'neutral80' }}
-        >
-          <UserSettingsButtonContents {...{ context, accountData, web3Context }} />
-        </Button>
-        {vaultFormToggleTitle && (
-          <Button variant="menuButton" sx={{ px: 3 }} onClick={() => setVaultFormOpened(true)}>
-            <Box>{vaultFormToggleTitle}</Box>
-          </Button>
-        )}
-      </Flex>
-      <MobileSidePanelPortal>
-        <Box
-          sx={{
-            display: 'block',
-            position: 'fixed',
-            left: 0,
-            right: 0,
-            bottom: 0,
-            transition: '0.3s transform ease-in-out',
-            transform: `translateY(${opened ? '0' : '100'}%)`,
-            bg: 'neutral10',
-            p: 3,
-            pt: 0,
-            zIndex: 'modal',
-            boxShadow: 'bottomSheet',
-            borderTopLeftRadius: 'large',
-            borderTopRightRadius: 'large',
-          }}
-          ref={componentRef}
-        >
-          <Box
-            sx={{
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'flex-end',
-              pt: 2,
-            }}
-          >
-            <ModalCloseIcon
-              close={() => setOpened(false)}
-              sx={{ top: 0, right: 0, color: 'primary100', position: 'relative' }}
-              size={3}
-            />
-          </Box>
-          <Card variant="vaultFormContainer" sx={{ p: 2 }}>
-            <UserSettings />
-          </Card>
-        </Box>
-      </MobileSidePanelPortal>
-    </>
-  )
-}
-
 function navLinkColor(isActive: boolean) {
   return isActive ? 'primary100' : 'neutral80'
 }
@@ -501,7 +416,7 @@ function ConnectedHeader() {
               />
               <MobileMenu />
             </Flex>
-            <MobileSettings />
+            <WalletPanelMobile />
           </BasicHeader>
         </Box>
       )}
